Ecommerce-backend/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
